# chore: use set to check supported algorithms